Fixed a crash. Individual compute unit results are no longer stored as the permutation can be of infinite size, resulting in a very large list of ComputeUnitResult's... Refactored code to support pruning of top X results and keeping these instead.
git-svn-id: svn://192.168.1.132/autoStock/autoStock@565 33c8eb2e-3fb0-4f7a-9790-3894f54e5fb4

+++ src/com/autoStock/cluster/ComputeUnitForBacktest.java
@@ -19,9 +19,9 @@ public class ComputeUnitForBacktest {
 	public Date dateStart;
 	public Date dateEnd;
 	public ArrayList<ArrayList<AdjustmentOfPortable>> listOfAdjustment = new ArrayList<ArrayList<AdjustmentOfPortable>>();
-	public int requestId;
+	public long requestId;
 	
-	public ComputeUnitForBacktest(int requestId, ArrayList<ArrayList<AdjustmentOfPortable>> listOfAdjustment, Exchange exchange, ArrayList<String> listOfSymbols, Date dateStart, Date dateEnd) {
+	public ComputeUnitForBacktest(long requestId, ArrayList<ArrayList<AdjustmentOfPortable>> listOfAdjustment, Exchange exchange, ArrayList<String> listOfSymbols, Date dateStart, Date dateEnd) {
 		this.requestId = requestId;
 		this.exchange = exchange;
 		this.listOfSymbols = listOfSymbols;
